feat: Add option to set a default allowed_mentions
For years, there has been such an option to enable the disableEveryone
option globally.

disableEveryone is hack that was created when Discord didn't yet provide
any option to prevent unexpected ping.

We now have a first party option to prevent unexpected pings,
allowed_mentions, but adoption is complicated because it must be set on
each individual message.

This commit allows lib users to set a default value for allowed_mentions
that will be used if not explicitly overriden when posting a message.