deprecate: Deprecate disableEveryone option
disableEveryone is hack that was created when Discord didn't yet provide
any option to prevent unexpected ping.

We now have a first party option to prevent unexpected pings,
allowed_mentions, which is better and less prone to failure.

There is no actual reason to continue to use disableEveryone. Therefore,
this commit marks the option as deprecated, and emits a deprecation
warning when the internal replaceEveryone function is called for the
first time.